+++ src/Speckle.Sdk.Artifacts.Harness/GraphArtifactProducer.cs
@@ -1,33 +1,781 @@
+using System.Collections;
 using System.Globalization;
 using Speckle.DoubleNumerics;
 using Speckle.Objects.Data;
+using Speckle.Objects.Geometry;
 using Speckle.Objects.Other;
 using Speckle.Objects.Utils;
+using Speckle.Sdk.Common;
 using Speckle.Sdk.Models;
 using Speckle.Sdk.Models.Collections;
+using Speckle.Sdk.Models.GraphTraversal;
 using Speckle.Sdk.Models.Instances;
 using Speckle.Sdk.Models.Proxies;
 
 namespace Speckle.Sdk.Artifacts.Harness;
 
 /// <summary>
-/// FIRST-CUT migration driver: walks a *stored* Speckle <see cref="Base"/> graph (the kind a connector
-/// sent — collections + atomic objects with displayValue + root-collection proxy arrays) and drives the
-/// validated <see cref="ObjectsArtifactPipeline"/> to emit the 3-artefact bundle (geometries + eav + envelope).
+/// Migrates a stored Speckle <see cref="Base"/> graph into the artefact bundle.
 ///
-/// This is the analogue of ODA's <c>RevitModelExtractor.EmitObject</c> + the post-loop value-node emission,
-/// but reading a deserialized graph instead of live ODA elements. It lives in the harness while we tune
-/// derivation correctness against real NDJSONs; promote to <c>src/Speckle.Objects/Utils</c> once proven.
-///
-/// KNOWN OPEN KEYING CHOICES (the "settle at implementation" items from topology-envelope-SOT §7) — first
-/// cut picks the ODA/BIM-consistent option and reports mismatches so we can validate against real data:
-///   • definition.objects[]      → geometry namespace (InternGeometryId)
-///   • material.objects[]        → geometry namespace (per-mesh; CAD is object-grained — flagged in stats)
-///   • color.objects[]           → object namespace
-///   • level.objects[]           → object namespace
+/// Discovery runs off <see cref="DefaultTraversal"/>, which yields exactly the two node kinds we migrate —
+/// every <see cref="Collection"/> and every ATOMIC object — depth-first, parent before child. Traversal only
+/// descends the <c>elements</c> lineage, so it halts at an atomic object: its <c>displayValue</c> and data
+/// properties are handled here, never walked into.
 /// </summary>
-internal static class GraphArtifactProducer
+internal sealed class GraphArtifactProducer(ObjectsArtifactPipeline pipeline) : IDisposable
 {
+  public void Dispose() => pipeline?.Dispose();
+
+  private readonly Stats _stats = new();
+
+  private readonly Dictionary<string, int> _collectionMap = new(StringComparer.Ordinal); // collection appId → node K
+  private readonly Dictionary<string, int> _objectKMap = new(StringComparer.Ordinal); // object id → object K, for SUBELEMENT hosts
+
+  // appIds actually emitted — lets proxy refs to absent targets be skipped rather than minting phantom Ks.
+  private readonly HashSet<string> _seenObjectAppIds = new(StringComparer.Ordinal);
+  private readonly HashSet<string> _seenGeometryAppIds = new(StringComparer.Ordinal);
+
+  // INSTANCE-node K by appId, shared between atomic instance leaves and nested-instance definition members.
+  private readonly Dictionary<string, int> _instanceNodeByAppId = new(StringComparer.Ordinal);
+
+  // object/geometry appId → its appearance targets ("g:<geomAppId>" mesh | "o:<objAppId>" instance object)
+  private readonly Dictionary<string, List<string>> _objectDisplayGeomKeys = new(StringComparer.Ordinal);
+
+  // geometry appId → material embedded on the mesh (v2-style sends).
+  private readonly Dictionary<string, RenderMaterial> _embeddedMaterialByGeom = new(StringComparer.Ordinal);
+
+  // definition-member appId → its raw-solid geometry appId ("<appId>:solid")
+  private readonly Dictionary<string, string> _defMemberSolidKey = new(StringComparer.Ordinal);
+
+  // pre v2.13 commits did not use Collection objects.
+  private bool _isv2PreCollections;
+
+  // synthetic collection K by cumulative property path ("Level 1/Walls") — dedup + parent-chain building (v2 mode).
+  private readonly Dictionary<string, int> _v2CollByPath = new(StringComparer.Ordinal);
+
+  /// <summary>Walks <paramref name="root"/>, drives the pipeline, and <see cref="ObjectsArtifactPipeline.Complete"/>s
+  /// it. Returns the run stats.</summary>
+  public Stats Produce(Base root)
+  {
+    _isv2PreCollections = root is not Collection;
+
+    IReadOnlySet<string> defSourceAppIds = GetDefinitionAppIds(root);
+
+    var traversal = DefaultTraversal.CreateTraversalFunc();
+    foreach (var tc in traversal.Traverse(root))
+    {
+      Base current = tc.Current;
+      TraversalContext? parent = tc.Parent;
+
+      if (parent is null)
+      {
+        continue;
+      }
+
+      if (current is Collection col)
+      {
+        // Parent is a Collection except at the root, which is skipped above and never mapped → top-level (null).
+        int? parentK = _collectionMap.TryGetValue(Aid(parent.Current), out var pk) ? pk : null;
+        var k = pipeline.AddCollection(CollectionKey(col), col.name, parentK, CollectionSubtype(col));
+        _collectionMap[Aid(col)] = k;
+        _stats.Collections++;
+        continue;
+      }
+
+      if (defSourceAppIds.Contains(Aid(current)))
+      {
+        EmitDefinitionMember(current);
+        continue;
+      }
+
+      var objK = EmitObject(current);
+      _objectKMap[current.id.NotNull()] = objK;
+      if (_isv2PreCollections)
+      {
+        EmitV2HierarchyEdge(tc, objK);
+      }
+      else
+      {
+        EmitHierarchyEdge(tc, objK);
+      }
+    }
+
+    var layerGeomKeys = BuildLayerGeomKeys(root, out var layerDepth);
+
+    EmitProxies(root, layerGeomKeys, layerDepth);
+
+    _stats.Geometries = _seenGeometryAppIds.Count;
+    pipeline.Complete();
+    return _stats;
+  }
+
+  // ── hierarchy edges ─────────────────────────────────────────────────────────────────
+
+  // Every atomic object belongs to its nearest ancestor collection (IN_COLLECTION), regardless of what sits
+  // between them. If an object host (e.g. a DataObject carrying `elements`) is one of those in-between nodes,
+  // it also gets a host→hosted SUBELEMENT to the nearest such host.
+  private void EmitHierarchyEdge(TraversalContext tc, int objK)
+  {
+    var hostFound = false;
+    for (var p = tc.Parent; p is not null; p = p.Parent)
+    {
+      var pc = p.Current;
+      if (pc is Collection)
+      {
+        if (_collectionMap.TryGetValue(Aid(pc), out var ck))
+        {
+          pipeline.InCollection(objK, ck, 0);
+          _stats.InCollectionEdges++;
+        }
+        return; // nearest collection reached; membership resolved
+      }
+      if (!hostFound && _objectKMap.TryGetValue(pc.id.NotNull(), out var hostK))
+      {
+        pipeline.Subelement(hostK, objK, _stats.SubelementEdges++);
+        hostFound = true;
+      }
+      // keep walking up to reach the enclosing collection
+    }
+  }
+
+  // Backwards compat pre 2.13 Collections
+  private void EmitV2HierarchyEdge(TraversalContext tc, int objK)
+  {
+    if (tc.Parent is not null && _objectKMap.TryGetValue(tc.Parent.Current.id.NotNull(), out var hostK))
+    {
+      pipeline.Subelement(hostK, objK, _stats.SubelementEdges++);
+    }
+
+    int? parentK = null;
+    var path = "";
+    foreach (var name in tc.GetPropertyPath().Reverse()) // child→root, so reverse to build root→leaf
+    {
+      path = path.Length == 0 ? name : path + "/" + name;
+      if (!_v2CollByPath.TryGetValue(path, out var pathK))
+      {
+        pathK = pipeline.AddCollection(path, name.TrimStart('@'), parentK, "Layer");
+        _v2CollByPath[path] = pathK;
+        _stats.Collections++;
+      }
+      parentK = pathK;
+    }
+    var leafK = parentK; // null when the object sits directly on the root (no property path) → stays top-level
+
+    if (leafK is not { } ck)
+    {
+      throw new InvalidOperationException("Object has no collection");
+    }
+
+    pipeline.InCollection(objK, ck, 0);
+    _stats.InCollectionEdges++;
+  }
+
+  // ── object emission ─────────────────────────────────────────────────────────────────
+
+  private int EmitObject(Base obj)
+  {
+    var appId = Aid(obj);
+    var objK = pipeline.InternObject(appId);
+    if (!_seenObjectAppIds.Add(appId))
+    {
+      return objK; // shared reference, same K
+    }
+
+    _stats.Objects++;
+
+    var (props, rootScalars, typeKey) = ExtractProperties(obj);
+    pipeline.AddProperties(appId, props, rootScalars, typeKey);
+
+    if (obj is InstanceProxy ip)
+    {
+      var instK = ResolveInstanceNode(ip);
+      pipeline.DisplayInstance(objK, instK, 0);
+      _stats.DisplayInstanceEdges++;
+      _stats.InstanceAtomics++;
+      // ByBlock: the placed definition geometry is shared, so appearance rides on the instance object.
+      RecordObjectGeom(appId, "o:" + appId);
+      return objK;
+    }
+
+    // Lossless raw solid (Brep/Extrusion/SubD/SolidX, or a Rhino/Autocad host wrapper): link the native blob via
+    // the SOLID rel, in ADDITION to the display meshes below. Receive picks solid vs mesh via PreferSolids.
+    var rawEnc = TryReadRawEncoding(obj);
+    if (rawEnc is not null && IsMigratableSolidFormat(rawEnc.format) && EmitSolidBlob(appId, rawEnc) is int solidK)
+    {
+      pipeline.Solid(objK, solidK, 0);
+      _stats.Solids++;
+      // Deliberately NOT recorded in _objectDisplayGeomKeys — materials/colours bind to display meshes, not the solid.
+    }
+    // Checked before the raw-geometry case so a leaf that ships a display mesh (Brep/SubD, extrusions) encodes
+    // that mesh rather than its un-encodable self.
+    var displayValue = GetBaseList(obj, "displayValue").ToList();
+    if (displayValue.Count > 0)
+    {
+      // Element-level material fills gaps where a mesh carries none of its own.
+      var objMaterial = ReadEmbeddedMaterial(obj);
+      int ord = 0;
+      foreach (var item in displayValue)
+      {
+        if (item is InstanceProxy dip)
+        {
+          var dInstK = ResolveInstanceNode(dip);
+          pipeline.DisplayInstance(objK, dInstK, ord++);
+          _stats.DisplayInstanceEdges++;
+        }
+        else
+        {
+          var gAppId = Aid(item);
+          if (AddGeometry(gAppId, item))
+          {
+            pipeline.Display(objK, pipeline.InternGeometryId(gAppId), ord++);
+            _stats.DisplayEdges++;
+            RecordObjectGeom(appId, "g:" + gAppId);
+            if ((ReadEmbeddedMaterial(item) ?? objMaterial) is { } rm)
+            {
+              _embeddedMaterialByGeom.TryAdd(gAppId, rm);
+            }
+          }
+        }
+      }
+      return objK;
+    }
+
+    // A display-less raw-encoded object (e.g. a BrepX with an empty displayValue) has no SGEO encoding; its solid
+    // blob is already captured above, so don't fall through and feed it to SgeoEncoder (which would throw).
+    if (obj is IRawEncodedObject)
+    {
+      return objK;
+    }
+
+    if (IsGeometry(obj))
+    {
+      // The leaf is its own geometry; appId interns into both the object and geometry namespaces.
+      if (AddGeometry(appId, obj))
+      {
+        pipeline.Display(objK, pipeline.InternGeometryId(appId), 0);
+        _stats.DisplayEdges++;
+        _stats.MeshAtomics++;
+        RecordObjectGeom(appId, "g:" + appId);
+        if (ReadEmbeddedMaterial(obj) is { } rm)
+        {
+          _embeddedMaterialByGeom.TryAdd(appId, rm);
+        }
+      }
+    }
+
+    return objK;
+  }
+
+  private void RecordObjectGeom(string objAppId, string geomAppId)
+  {
+    if (!_objectDisplayGeomKeys.TryGetValue(objAppId, out var list))
+    {
+      list = new List<string>();
+      _objectDisplayGeomKeys[objAppId] = list;
+    }
+    list.Add(geomAppId);
+  }
+
+  // Idempotent by appId; also emits the placement's DEFINITION so DEFINES_INSTANCE can resolve.
+  private int ResolveInstanceNode(InstanceProxy ip)
+  {
+    var key = Aid(ip);
+    if (_instanceNodeByAppId.TryGetValue(key, out var existing))
+    {
+      return existing;
+    }
+    var defK = pipeline.AddDefinition(ip.definitionId, ip.definitionId);
+    var instK = pipeline.AddInstance(key, defK, Flatten(ip.transform), ip.units);
+    _instanceNodeByAppId[key] = instK;
+    return instK;
+  }
+
+  private bool AddGeometry(string appId, Base geometry)
+  {
+    pipeline.AddGeometry(appId, geometry);
+    _seenGeometryAppIds.Add(appId);
+    return true;
+  }
+
+  // Definition content: a nested instance → INSTANCE node (linked via DEFINES_INSTANCE); otherwise a geometry
+  // blob under the member's appId (linked via DEFINES), preferring a display mesh over an un-encodable parent.
+  private void EmitDefinitionMember(Base obj)
+  {
+    var appId = Aid(obj);
+
+    if (obj is InstanceProxy ip)
+    {
+      if (!_instanceNodeByAppId.ContainsKey(appId))
+      {
+        ResolveInstanceNode(ip);
+        _stats.DefinitionInstances++;
+      }
+      return;
+    }
+
+    if (!_seenGeometryAppIds.Add(appId))
+    {
+      return;
+    }
+
+    // Lossless raw solid for a definition member rides DEFINES (no standalone SOLID edge) alongside its display
+    // meshes — the EmitProxies DEFINES pass links it via _defMemberSolidKey (mirrors the Rhino connector).
+    var rawEnc = TryReadRawEncoding(obj);
+    var hasSolid =
+      rawEnc is not null && IsMigratableSolidFormat(rawEnc.format) && EmitSolidBlob(appId, rawEnc) is not null;
+    if (hasSolid)
+    {
+      _defMemberSolidKey[appId] = appId + ":solid";
+      _stats.DefinitionSolids++;
+    }
+
+    // Don't SGEO-encode a raw-encoded object itself (BrepX/…) — only a genuine display mesh or plain geometry leaf.
+    var geometry =
+      GetBaseList(obj, "displayValue").FirstOrDefault(d => d is not InstanceProxy)
+      ?? (obj is not IRawEncodedObject && IsGeometry(obj) ? obj : null);
+    if (geometry is null)
+    {
+      _seenGeometryAppIds.Remove(appId); // no SGEO mesh under appId; a solid-only member is linked via _defMemberSolidKey
+      if (!hasSolid)
+      {
+        _stats.SkippedDefines++;
+        if (_stats.Notes.Count < 20)
+        {
+          _stats.Notes.Add($"def member {appId} has no encodable geometry [{obj.speckle_type}]");
+        }
+      }
+      return;
+    }
+
+    pipeline.AddGeometry(appId, geometry);
+    _stats.DefinitionGeometries++;
+    if ((ReadEmbeddedMaterial(geometry) ?? ReadEmbeddedMaterial(obj)) is { } rm)
+    {
+      _embeddedMaterialByGeom.TryAdd(appId, rm);
+    }
+    // Shared across placements, so ByLayer colour must bind to this geometry-K, not flood from an instance.
+    _objectDisplayGeomKeys[appId] = new List<string> { "g:" + appId };
+  }
+
+  // ── proxy / value-node emission ─────────────────────────────────────────────────────
+
+  private void EmitProxies(
+    Base root,
+    Dictionary<string, List<string>> layerGeomKeys,
+    Dictionary<string, int> layerDepth
+  )
+  {
+    // A proxy ref resolved to its tagged appearance targets ("g:" mesh | "o:" instance object).
+    IReadOnlyList<string> DirectGeomKeys(string refAppId)
+    {
+      if (_objectDisplayGeomKeys.TryGetValue(refAppId, out var og))
+      {
+        return og;
+      }
+      if (_seenGeometryAppIds.Contains(refAppId))
+      {
+        return new[] { "g:" + refAppId };
+      }
+      return Array.Empty<string>();
+    }
+
+    // Binds value-nodes to geometry. A direct (mesh/object) ref claims a geometry; a Layer ref then fills only
+    // what's unclaimed, and among nested layers the deepest wins (else a parent layer floods its sub-layers).
+    Dictionary<string, int> BindWithPrecedence(List<(int nodeK, List<string> refs)> proxies, out int skipped)
+    {
+      var byGeom = new Dictionary<string, int>(StringComparer.Ordinal);
+      var skip = 0;
+      foreach (var (nodeK, refs) in proxies)
+      {
+        foreach (var r in refs)
+        {
+          var direct = DirectGeomKeys(r);
+          if (direct.Count == 0 && !layerGeomKeys.ContainsKey(r))
+          {
+            skip++;
+          }
+          foreach (var gk in direct)
+          {
+            byGeom.TryAdd(gk, nodeK);
+          }
+        }
+      }
+      var layerCandidate = new Dictionary<string, (int depth, int nodeK)>(StringComparer.Ordinal);
+      foreach (var (nodeK, refs) in proxies)
+      {
+        foreach (var r in refs)
+        {
+          if (!layerGeomKeys.TryGetValue(r, out var lg))
+          {
+            continue;
+          }
+          var depth = layerDepth.GetValueOrDefault(r, 0);
+          foreach (var gk in lg)
+          {
+            if (byGeom.ContainsKey(gk))
+            {
+              continue; // a direct ref already claimed it
+            }
+            if (!layerCandidate.TryGetValue(gk, out var cur) || depth > cur.depth)
+            {
+              layerCandidate[gk] = (depth, nodeK);
+            }
+          }
+        }
+      }
+      foreach (var (gk, (_, nodeK)) in layerCandidate)
+      {
+        byGeom[gk] = nodeK;
+      }
+      skipped = skip;
+      return byGeom;
+    }
+
+    foreach (var def in GetBaseList(root, "instanceDefinitionProxies"))
+    {
+      if (def is not InstanceDefinitionProxy idp)
+      {
+        continue;
+      }
+      var defK = pipeline.AddDefinition(DefinitionKey(idp), idp.name);
+      _stats.Definitions++;
+      int o = 0;
+      foreach (var memberAppId in idp.objects)
+      {
+        if (_instanceNodeByAppId.TryGetValue(memberAppId, out var instK))
+        {
+          pipeline.DefinesInstance(defK, instK, o++);
+          _stats.DefinesInstanceEdges++;
+        }
+        else
+        {
+          // Solid and display mesh both DEFINES under the same member ordinal (solid first) so receive can
+          // group them and prefer the solid.
+          var any = false;
+          if (_defMemberSolidKey.TryGetValue(memberAppId, out var solidKey))
+          {
+            pipeline.Defines(defK, pipeline.InternGeometryId(solidKey), o);
+            _stats.DefinesEdges++;
+            any = true;
+          }
+          if (_seenGeometryAppIds.Contains(memberAppId))
+          {
+            pipeline.Defines(defK, pipeline.InternGeometryId(memberAppId), o);
+            _stats.DefinesEdges++;
+            any = true;
+          }
+          if (any)
+          {
+            o++;
+          }
+          else
+          {
+            _stats.SkippedDefines++;
+          }
+        }
+      }
+    }
+
+    var matProxies = new List<(int, List<string>)>();
+    var placeholderMatKs = new HashSet<int>();
+    foreach (var mat in GetBaseList(root, "renderMaterialProxies"))
+    {
+      if (mat is not RenderMaterialProxy rmp)
+      {
+        continue;
+      }
+      var v = rmp.value;
+      var matK = pipeline.AddMaterial(MaterialKey(rmp), v.diffuse, v.opacity, v.metalness, v.roughness);
+      matProxies.Add((matK, rmp.objects));
+      _stats.Materials++;
+      // Pure-black diffuse is the CAD "no material / ByLayer" placeholder; it must yield to a real colour.
+      if ((v.diffuse & 0xFFFFFF) == 0)
+      {
+        placeholderMatKs.Add(matK);
+      }
+    }
+    var matBindings = BindWithPrecedence(matProxies, out var matSkipped);
+    _stats.SkippedMaterial += matSkipped;
+
+    // Fallback for sends with no root material proxies: bind each mesh's embedded material directly (a proxy
+    // claim wins). Deduped by material identity. An embedded black is explicit, not the CAD placeholder.
+    var embeddedMatKs = new Dictionary<string, int>(StringComparer.Ordinal);
+    foreach (var (gAppId, rm) in _embeddedMaterialByGeom)
+    {
+      var gk = "g:" + gAppId;
+      if (matBindings.ContainsKey(gk))
+      {
+        continue;
+      }
+      var key = rm.applicationId ?? "mat:" + (rm.id ?? rm.diffuse.ToString(CultureInfo.InvariantCulture));
+      if (!embeddedMatKs.TryGetValue(key, out var matK))
+      {
+        matK = pipeline.AddMaterial(key, rm.diffuse, rm.opacity, rm.metalness, rm.roughness);
+        embeddedMatKs[key] = matK;
+        _stats.Materials++;
+      }
+      matBindings[gk] = matK;
+    }
+
+    var colProxies = new List<(int, List<string>)>();
+    foreach (var col in GetBaseList(root, "colorProxies"))
+    {
+      if (col is not ColorProxy cp)
+      {
+        continue;
+      }
+      colProxies.Add((pipeline.AddColor(cp.value), cp.objects));
+      _stats.Colors++;
+    }
+    var colBindings = BindWithPrecedence(colProxies, out var colSkipped);
+    _stats.SkippedColor += colSkipped;
+
+    foreach (var (target, matK) in matBindings)
+    {
+      if (target[0] != 'g')
+      {
+        continue; // instances carry no per-mesh material
+      }
+      // Placeholder black yields to a real display colour on the same mesh.
+      if (placeholderMatKs.Contains(matK) && colBindings.ContainsKey(target))
+      {
+        continue;
+      }
+      pipeline.HasMaterial(pipeline.InternGeometryId(target[2..]), matK);
+      _stats.HasMaterialEdges++;
+    }
+    // Geometry-targeted only: an object-target would collide with a geometry-target at the same numeric K.
+    foreach (var (target, colK) in colBindings)
+    {
+      if (target[0] == 'o')
+      {
+        continue;
+      }
+      pipeline.HasColor(pipeline.InternGeometryId(target[2..]), colK);
+      _stats.HasColorEdges++;
+    }
+
+    // Refs to elements absent from the export are skipped.
+    foreach (var lvl in GetBaseList(root, "levelProxies"))
+    {
+      if (lvl is not LevelProxy lp)
+      {
+        continue;
+      }
+      var name = lp.value.name;
+      var elevation = lp.value["elevation"] is double d ? d : 0.0; // dynamic member on the level DataObject
+      var lvlK = pipeline.AddLevel(LevelKey(lp, name), name, elevation);
+      _stats.Levels++;
+      foreach (var objAppId in lp.objects)
+      {
+        if (!_seenObjectAppIds.Contains(objAppId))
+        {
+          _stats.SkippedLevel++;
+          continue;
+        }
+        pipeline.OnLevel(pipeline.InternObject(objAppId), lvlK);
+        _stats.OnLevelEdges++;
+      }
+    }
+  }
+
+  // ── layer geometry resolution (ByLayer material/colour) ─────────────────────────────
+
+  // collection appId → the display geometry of every object beneath it, with each collection's depth — so a
+  // ByLayer proxy (which references a layer) can bind to that layer's meshes.
+  private Dictionary<string, List<string>> BuildLayerGeomKeys(Base root, out Dictionary<string, int> layerDepth)
+  {
+    var result = new Dictionary<string, List<string>>(StringComparer.Ordinal);
+    var depths = new Dictionary<string, int>(StringComparer.Ordinal);
+    void Walk(Base node, int depth)
+    {
+      foreach (var child in GetBaseList(node, "elements"))
+      {
+        if (child is not Collection)
+        {
+          continue;
+        }
+        var geoms = new List<string>();
+        CollectDescendantGeom(child, geoms);
+        if (geoms.Count > 0)
+        {
+          result[Aid(child)] = geoms;
+          depths[Aid(child)] = depth;
+        }
+        Walk(child, depth + 1);
+      }
+    }
+    Walk(root, 0);
+    layerDepth = depths;
+    return result;
+  }
+
+  private void CollectDescendantGeom(Base collection, List<string> acc)
+  {
+    foreach (var child in GetBaseList(collection, "elements"))
+    {
+      if (child is Collection)
+      {
+        CollectDescendantGeom(child, acc);
+      }
+      else if (_objectDisplayGeomKeys.TryGetValue(Aid(child), out var gks))
+      {
+        acc.AddRange(gks);
+      }
+    }
+  }
+
+  // ── property extraction ─────────────────────────────────────────────────────────────
+
+  private static (
+    IReadOnlyDictionary<string, object?> props,
+    IEnumerable<KeyValuePair<string, object?>> rootScalars,
+    string? typeKey
+  ) ExtractProperties(Base obj)
+  {
+    IReadOnlyDictionary<string, object?> props = obj is DataObject dobj
+      ? dobj.properties
+      : obj.GetMembers(DynamicBaseMemberType.Instance | DynamicBaseMemberType.Dynamic);
+
+    // `level` is the level NAME and lives at the top level (not under properties), so it must be listed here.
+    var rootScalars = new List<KeyValuePair<string, object?>>
+    {
+      new("speckle_type", obj.speckle_type),
+      new("name", obj["name"]),
+      new("units", obj["units"]),
+      new("category", obj["category"]),
+      new("family", obj["family"]),
+      new("type", obj["type"]),
+      new("level", obj["level"]),
+    };
+
+    var typeKey = obj["typeId"] as string ?? (props.TryGetValue("typeId", out var tk) ? tk as string : null);
+
+    return (props, rootScalars, typeKey);
+  }
+
+  // ── keys (applicationId-keyed; null → stable per-object key) ─────────────────────────
+
+  private static string Aid(Base b) => b.applicationId ?? "spk:" + b.id;
+
+  private static string CollectionKey(Collection col) => col.applicationId ?? "coll:" + col.id;
+
+  private static string DefinitionKey(InstanceDefinitionProxy idp) => idp.applicationId ?? idp.name;
+
+  private static string MaterialKey(RenderMaterialProxy rmp) =>
+    rmp.applicationId ?? rmp.value.applicationId ?? "mat:" + rmp.value.diffuse.ToString(CultureInfo.InvariantCulture);
+
+  private static string LevelKey(Base lvl, string? name) => lvl.applicationId ?? "lvl:" + (name ?? lvl.id);
+
+  private static string CollectionSubtype(Collection col)
+  {
+#pragma warning disable CS0618 // Type or member is obsolete
+    var ct = col.collectionType;
+#pragma warning restore CS0618 // Type or member is obsolete
+    return string.IsNullOrEmpty(ct) ? col.speckle_type.Split('.')[^1] : ct;
+  }
+
+  // ── helpers ─────────────────────────────────────────────────────────────────────────
+
+  private HashSet<string> GetDefinitionAppIds(Base root)
+  {
+    var set = new HashSet<string>(StringComparer.Ordinal);
+    foreach (var def in GetBaseList(root, "instanceDefinitionProxies"))
+    {
+      if (def is not InstanceDefinitionProxy idp)
+      {
+        throw new InvalidOperationException(
+          $"Encountered a {def.GetType()} where expected a {typeof(InstanceDefinitionProxy)}"
+        );
+      }
+
+      foreach (var appId in idp.objects)
+      {
+        set.Add(appId);
+      }
+    }
+    return set;
+  }
+
+  private static bool IsGeometry(Base b) => b.speckle_type.StartsWith("Objects.Geometry.", StringComparison.Ordinal);
+
+  private static RenderMaterial? ReadEmbeddedMaterial(Base host) =>
+    (host["renderMaterial"] ?? host["@renderMaterial"]) as RenderMaterial;
+
+  // ── raw-encoded solids (SOLID rel) ───────────────────────────────────────────────────
+
+  // 3dm (Rhino) and sat (Autocad) are the native solid formats we migrate; others (e.g. dwg) are skipped.
+  private static bool IsMigratableSolidFormat(string? format) =>
+    format is RawEncodingFormats.RHINO_3DM or RawEncodingFormats.ACAD_SAT;
+
+  // Reads the lossless raw encoding off a raw-encoded geometry (encodedValue) or a host wrapper (rawEncoding).
+  // Typed casts only — the v3 graph deserializes into these registered SDK types.
+  private static RawEncoding? TryReadRawEncoding(Base obj) =>
+    obj switch
+    {
+      IRawEncodedObject r => r.encodedValue,
+      RhinoObject ro => ro.rawEncoding,
+      AutocadObject ao => ao.rawEncoding,
+      _ => null,
+    };
+
+  // base64-decodes the blob and stores it under "<objAppId>:solid" as a raw (non-SGEO) geometry, returning its K.
+  // A malformed blob is logged and skipped (null) rather than failing the whole migration run.
+  private int? EmitSolidBlob(string objAppId, RawEncoding enc)
+  {
+    byte[] bytes;
+    try
+    {
+      bytes = Convert.FromBase64String(enc.contents);
+      return pipeline.AddRawGeometry(objAppId + ":solid", bytes, enc.format);
+    }
+    catch (FormatException ex)
+    {
+      throw new InvalidOperationException($"Malformed solid {objAppId}", ex);
+    }
+  }
+
+  // A detached list may sit under the typed key or the `@`-prefixed dynamic key; take the first non-empty one.
+  private static IEnumerable<Base> GetBaseList(Base b, string key)
+  {
+    var raw = NonEmpty(b[key]) ?? NonEmpty(b["@" + key]);
+    if (raw is IEnumerable seq and not string)
+    {
+      foreach (var item in seq)
+      {
+        if (item is Base bs)
+        {
+          yield return bs;
+        }
+      }
+    }
+  }
+
+  private static object? NonEmpty(object? v) => v is ICollection c && c.Count == 0 ? null : v;
+
+  private static double[] Flatten(Matrix4x4 m) =>
+    new[]
+    {
+      m.M11,
+      m.M12,
+      m.M13,
+      m.M14,
+      m.M21,
+      m.M22,
+      m.M23,
+      m.M24,
+      m.M31,
+      m.M32,
+      m.M33,
+      m.M34,
+      m.M41,
+      m.M42,
+      m.M43,
+      m.M44,
+    };
+
+  // ── stats ───────────────────────────────────────────────────────────────────────────
+
   internal sealed class Stats
   {
     public int Objects;
@@ -51,8 +799,11 @@ internal static class GraphArtifactProducer
     public int MeshAtomics;
     public int InstanceAtomics;
 
-    // Proxy refs skipped because the target appId is absent from the graph (e.g. level membership to a
-    // Revit Room filtered out of the export). Skipped rather than minting a phantom object/geometry K.
+    // Raw (non-SGEO) native solid blobs: SOLID edges on atomic objects, DEFINES-linked solids on def members.
+    public int Solids;
+    public int DefinitionSolids;
+
+    // Proxy refs whose target appId isn't in the graph — skipped rather than minting a phantom K.
     public int SkippedDefines;
     public int SkippedMaterial;
     public int SkippedColor;
@@ -63,900 +814,10 @@ internal static class GraphArtifactProducer
     public override string ToString() =>
       $"""
         objects={Objects} (meshAtomic={MeshAtomics} instAtomic={InstanceAtomics})  geometries={Geometries} (defGeom={DefinitionGeometries})
-        edges: DISPLAY={DisplayEdges} DISPLAY_INSTANCE={DisplayInstanceEdges} SUBELEMENT={SubelementEdges}
+        edges: DISPLAY={DisplayEdges} DISPLAY_INSTANCE={DisplayInstanceEdges} SUBELEMENT={SubelementEdges} SOLID={Solids} (defSolid={DefinitionSolids})
                DEFINES={DefinesEdges} DEFINES_INSTANCE={DefinesInstanceEdges} HAS_MATERIAL={HasMaterialEdges} HAS_COLOR={HasColorEdges} ON_LEVEL={OnLevelEdges} IN_COLLECTION={InCollectionEdges}
         nodes: DEFINITION={Definitions} INSTANCE(def)={DefinitionInstances} MATERIAL={Materials} COLOR={Colors} LEVEL={Levels} COLLECTION={Collections}
         skipped (ref not in graph): {SkippedDangling}  (DEFINES={SkippedDefines} HAS_MATERIAL={SkippedMaterial} HAS_COLOR={SkippedColor} ON_LEVEL={SkippedLevel})
         """;
   }
-
-  public static Stats Produce(Base root, string outDir, string baseName)
-  {
-    Directory.CreateDirectory(outDir);
-    var stats = new Stats();
-    using var pipeline = new ObjectsArtifactPipeline(outDir, baseName);
-
-    // appIds we have actually emitted as objects / geometry — to detect dangling proxy refs.
-    var seenObjectAppIds = new HashSet<string>(StringComparer.Ordinal);
-    var seenGeometryAppIds = new HashSet<string>(StringComparer.Ordinal);
-    // INSTANCE-node K by appId — for both atomic instance leaves and nested-instance definition members,
-    // so the DEFINES_INSTANCE (rel 9) edges from a definition to a nested instance resolve to the node.
-    var instanceNodeByAppId = new Dictionary<string, int>(StringComparer.Ordinal);
-    // object appId → the geometry appIds of its DISPLAY meshes — lets OBJECT-grained material/colour proxy
-    // refs (which name the DataObject, not its mesh — Navis/CAD) bind to the object's display geometry.
-    var objectDisplayGeomKeys = new Dictionary<string, List<string>>(StringComparer.Ordinal);
-    // geometry appId → the mesh's embedded RenderMaterial. Old-style sends (pre-proxy Revit/Rhino) attach
-    // the material inline on each display mesh (or its parent element) instead of publishing root-level
-    // renderMaterialProxies — collected during the walk, folded into the material bindings in EmitProxies
-    // (a root-proxy claim wins over this fallback).
-    var embeddedMaterialByGeom = new Dictionary<string, RenderMaterial>(StringComparer.Ordinal);
-
-    // Instance-definition source members: appIds in `instanceDefinitionProxies[].objects[]`. These are the
-    // definition's content (rendered via instance placements, never directly), NOT atomic scene objects —
-    // route each to GEOMETRY (raw-mesh member) or to an INSTANCE node (nested-instance member), so
-    // DEFINES/DEFINES_INSTANCE + HAS_MATERIAL resolve. Mirrors ODA's DefinitionGeometries.
-    var defSourceAppIds = CollectDefinitionSourceAppIds(root);
-
-    // Scene-tree containers (Rhino/CAD layers, ETABS story/category, Tekla category, …): one COLLECTION node
-    // per source Collection (root excluded), parent-chained, so the loader can rebuild the source hierarchy
-    // that flattening would otherwise lose. collectionMap: collection appId → COLLECTION node-K.
-    var collectionMap = BuildCollectionNodes(root, pipeline, stats);
-
-    // 1) Walk the collection tree. Definition members → geometry/instance-node; every other leaf is an
-    //    atomic object — type-agnostically (SOT §1): DataObject, raw mesh, or bare instance.
-    foreach (var (obj, parent) in TraverseAtomics(root))
-    {
-      if (defSourceAppIds.Contains(Aid(obj)))
-      {
-        EmitDefinitionMember(
-          pipeline,
-          obj,
-          stats,
-          seenGeometryAppIds,
-          instanceNodeByAppId,
-          objectDisplayGeomKeys,
-          embeddedMaterialByGeom
-        );
-        continue;
-      }
-
-      var objK = EmitObject(
-        pipeline,
-        obj,
-        stats,
-        seenObjectAppIds,
-        seenGeometryAppIds,
-        instanceNodeByAppId,
-        objectDisplayGeomKeys,
-        embeddedMaterialByGeom
-      );
-
-      // IN_COLLECTION: the object's direct membership in its enclosing scene-tree container (null parent or a
-      // non-collection parent — e.g. an object directly under the root — gets no edge).
-      if (parent is Collection && collectionMap.TryGetValue(Aid(parent), out var collK))
-      {
-        pipeline.InCollection(objK, collK, 0);
-        stats.InCollectionEdges++;
-      }
-
-      // host→hosted nesting: a non-collection object carrying `elements` (e.g. curtain wall → panels).
-      foreach (var child in GetBaseList(obj, "elements"))
-      {
-        if (child is Collection || defSourceAppIds.Contains(Aid(child)))
-        {
-          continue;
-        }
-        var childK = EmitObject(
-          pipeline,
-          child,
-          stats,
-          seenObjectAppIds,
-          seenGeometryAppIds,
-          instanceNodeByAppId,
-          objectDisplayGeomKeys,
-          embeddedMaterialByGeom
-        );
-        pipeline.Subelement(objK, childK, stats.SubelementEdges++);
-      }
-    }
-
-    // layer/collection appId → the geometry of all its descendant objects — for ByLayer material/colour
-    // (CAD proxies reference a Layer; resolve the walk-up at PRODUCE so the consumer stays flat).
-    var layerGeomKeys = BuildLayerGeomKeys(root, objectDisplayGeomKeys, out var layerDepth);
-
-    // 2) Value-nodes + their edges, from the root collection's proxy arrays.
-    EmitProxies(
-      pipeline,
-      root,
-      stats,
-      seenObjectAppIds,
-      seenGeometryAppIds,
-      instanceNodeByAppId,
-      objectDisplayGeomKeys,
-      layerGeomKeys,
-      layerDepth,
-      embeddedMaterialByGeom
-    );
-
-    stats.Geometries = seenGeometryAppIds.Count;
-
-    pipeline.Complete();
-    return stats;
-  }
-
-  // ── object emission ──────────────────────────────────────────────────────────────
-
-  private static int EmitObject(
-    ObjectsArtifactPipeline pipeline,
-    Base obj,
-    Stats stats,
-    HashSet<string> seenObjectAppIds,
-    HashSet<string> seenGeometryAppIds,
-    Dictionary<string, int> instanceNodeByAppId,
-    Dictionary<string, List<string>> objectDisplayGeomKeys,
-    Dictionary<string, RenderMaterial> embeddedMaterialByGeom
-  )
-  {
-    var appId = Aid(obj);
-    var objK = pipeline.InternObject(appId);
-    if (!seenObjectAppIds.Add(appId))
-    {
-      return objK; // already emitted (shared reference) — same K, idempotent intern
-    }
-
-    stats.Objects++;
-
-    var (props, rootScalars, typeKey) = ExtractProperties(obj);
-    pipeline.AddProperties(appId, props, rootScalars, typeKey);
-
-    // SOT §1 — atomic is type-agnostic; dispatch on the leaf shape:
-    if (obj is InstanceProxy ip)
-    {
-      // instance-atomic: the leaf IS a placement → object DISPLAY_INSTANCEs its own INSTANCE node.
-      var instK = ResolveInstanceNode(pipeline, ip, instanceNodeByAppId);
-      pipeline.DisplayInstance(objK, instK, 0);
-      stats.DisplayInstanceEdges++;
-      stats.InstanceAtomics++;
-      // Appearance target is the OBJECT itself (ByBlock): the placed definition geometry is shared, so a
-      // ByLayer colour must ride on the instance object — the consumer applies it during instance placement.
-      RecordObjectGeom(objectDisplayGeomKeys, appId, "o:" + appId);
-      return objK;
-    }
-
-    // A leaf that carries displayValue references its geometry there — covers DataObjects AND geometry
-    // leaves that aren't directly SGEO-encodable but ship a display mesh (Rhino BrepX/SubD, extrusions…).
-    // Checked BEFORE the raw-geometry case so we encode the display mesh, not the un-encodable parent.
-    var displayValue = GetBaseList(obj, "displayValue").ToList();
-    if (displayValue.Count > 0)
-    {
-      // Element-level embedded material: some old-style hosts (Revit walls/openings) carry the material on
-      // the ELEMENT while their meshes carry their own — the mesh's wins, the element's fills the gaps.
-      var objMaterial = ReadEmbeddedMaterial(obj);
-      int ord = 0;
-      foreach (var item in displayValue)
-      {
-        if (item is InstanceProxy dip)
-        {
-          var dInstK = ResolveInstanceNode(pipeline, dip, instanceNodeByAppId);
-          pipeline.DisplayInstance(objK, dInstK, ord++);
-          stats.DisplayInstanceEdges++;
-        }
-        else
-        {
-          var gAppId = GeometryKey(item);
-          if (TryAddGeometry(pipeline, gAppId, item, seenGeometryAppIds))
-          {
-            pipeline.Display(objK, pipeline.InternGeometryId(gAppId), ord++);
-            stats.DisplayEdges++;
-            RecordObjectGeom(objectDisplayGeomKeys, appId, "g:" + gAppId);
-            if ((ReadEmbeddedMaterial(item) ?? objMaterial) is { } rm)
-            {
-              embeddedMaterialByGeom.TryAdd(gAppId, rm);
-            }
-          }
-        }
-      }
-      return objK;
-    }
-
-    if (IsGeometry(obj))
-    {
-      // mesh-atomic: the leaf IS the geometry → DISPLAY(object-K → its own geometry-K). The appId interns
-      // into BOTH the object and geometry namespaces (separate counters, no collision).
-      if (TryAddGeometry(pipeline, appId, obj, seenGeometryAppIds))
-      {
-        pipeline.Display(objK, pipeline.InternGeometryId(appId), 0);
-        stats.DisplayEdges++;
-        stats.MeshAtomics++;
-        RecordObjectGeom(objectDisplayGeomKeys, appId, "g:" + appId);
-        if (ReadEmbeddedMaterial(obj) is { } rm)
-        {
-          embeddedMaterialByGeom.TryAdd(appId, rm);
-        }
-      }
-    }
-
-    // else: a non-geometry leaf with no displayValue — an eav object with no renderable geometry.
-    return objK;
-  }
-
-  // Records that object `objAppId` renders geometry `geomAppId` — so an OBJECT-grained material/colour ref
-  // (proxies that name the DataObject, not its mesh) can bind to the object's display meshes.
-  private static void RecordObjectGeom(Dictionary<string, List<string>> map, string objAppId, string geomAppId)
-  {
-    if (!map.TryGetValue(objAppId, out var list))
-    {
-      list = new List<string>();
-      map[objAppId] = list;
-    }
-    list.Add(geomAppId);
-  }
-
-  // Intern/emit an INSTANCE node for a placement (idempotent by appId), recording its K so DEFINES_INSTANCE
-  // edges to a nested instance can resolve. AddDefinition links it to its own DEFINITION via def_ref.
-  private static int ResolveInstanceNode(
-    ObjectsArtifactPipeline pipeline,
-    InstanceProxy ip,
-    Dictionary<string, int> instanceNodeByAppId
-  )
-  {
-    var key = Aid(ip);
-    if (instanceNodeByAppId.TryGetValue(key, out var existing))
-    {
-      return existing;
-    }
-    var defK = pipeline.AddDefinition(ip.definitionId, ip.definitionId);
-    var instK = pipeline.AddInstance(key, defK, Flatten(ip.transform), ip.units);
-    instanceNodeByAppId[key] = instK;
-    return instK;
-  }
-
-  // Encode a geometry blob under `appId` (once). Returns false (and counts) on encode failure.
-  private static bool TryAddGeometry(
-    ObjectsArtifactPipeline pipeline,
-    string appId,
-    Base geometry,
-    HashSet<string> seenGeometryAppIds
-  )
-  {
-    pipeline.AddGeometry(appId, geometry);
-    seenGeometryAppIds.Add(appId);
-    return true;
-  }
-
-  // The union of all appIds referenced by the root's instance-definition proxies — i.e. the definition
-  // source geometry that must live in the geometry namespace, not the object set.
-  private static HashSet<string> CollectDefinitionSourceAppIds(Base root)
-  {
-    var set = new HashSet<string>(StringComparer.Ordinal);
-    foreach (var def in GetBaseList(root, "instanceDefinitionProxies"))
-    {
-      if (def is InstanceDefinitionProxy idp)
-      {
-        foreach (var appId in idp.objects)
-        {
-          set.Add(appId);
-        }
-      }
-    }
-    return set;
-  }
-
-  // A definition member (referenced by instanceDefinitionProxies[].objects[]) is definition CONTENT, not an
-  // atomic scene object. Route by shape: a raw mesh → geometry blob (linked later via DEFINES); a nested
-  // instance → its own INSTANCE node (linked via DEFINES_INSTANCE). A display-bearing member contributes its
-  // first display mesh under the member's appId.
-  private static void EmitDefinitionMember(
-    ObjectsArtifactPipeline pipeline,
-    Base obj,
-    Stats stats,
-    HashSet<string> seenGeometryAppIds,
-    Dictionary<string, int> instanceNodeByAppId,
-    Dictionary<string, List<string>> objectDisplayGeomKeys,
-    Dictionary<string, RenderMaterial> embeddedMaterialByGeom
-  )
-  {
-    var appId = Aid(obj);
-
-    if (obj is InstanceProxy ip)
-    {
-      // nested-instance member → INSTANCE node; the outer definition links it via DEFINES_INSTANCE.
-      if (!instanceNodeByAppId.ContainsKey(appId))
-      {
-        ResolveInstanceNode(pipeline, ip, instanceNodeByAppId);
-        stats.DefinitionInstances++;
-      }
-      return;
-    }
-
-    if (!seenGeometryAppIds.Add(appId))
-    {
-      return;
-    }
-    // Prefer a display mesh (encodable) over the parent geometry, which may be an un-encodable Brep/SubD.
-    var geometry =
-      GetBaseList(obj, "displayValue").FirstOrDefault(d => d is not InstanceProxy) ?? (IsGeometry(obj) ? obj : null);
-    if (geometry is null)
-    {
-      seenGeometryAppIds.Remove(appId);
-      stats.SkippedDefines++;
-      if (stats.Notes.Count < 20)
-      {
-        stats.Notes.Add($"def member {appId} has no encodable geometry [{obj.speckle_type}]");
-      }
-      return;
-    }
-
-    pipeline.AddGeometry(appId, geometry);
-    stats.DefinitionGeometries++;
-    if ((ReadEmbeddedMaterial(geometry) ?? ReadEmbeddedMaterial(obj)) is { } rm)
-    {
-      embeddedMaterialByGeom.TryAdd(appId, rm);
-    }
-    // Record the definition geometry under its own appId so the ByLayer walk (CollectDescendantGeom) can
-    // reach it: block content sits under its SOURCE layer Collection in the tree, and ByLayer colour must
-    // bind to this shared geometry-K (fixed across placements) — NOT flood from the instance's layer. The
-    // consumer's appearanceFor prefers colors.get(geomK) over colors.get(objK), so the source-layer colour
-    // wins over the instance's ByBlock fallback. (Tagged g: — materials/colours are geometry-targeted.)
-    objectDisplayGeomKeys[appId] = new List<string> { "g:" + appId };
-  }
-
-  // ── proxy / value-node emission ─────────────────────────────────────────────────
-
-  private static void EmitProxies(
-    ObjectsArtifactPipeline pipeline,
-    Base root,
-    Stats stats,
-    HashSet<string> seenObjectAppIds,
-    HashSet<string> seenGeometryAppIds,
-    Dictionary<string, int> instanceNodeByAppId,
-    Dictionary<string, List<string>> objectDisplayGeomKeys,
-    Dictionary<string, List<string>> layerGeomKeys,
-    Dictionary<string, int> layerDepth,
-    Dictionary<string, RenderMaterial> embeddedMaterialByGeom
-  )
-  {
-    // DIRECT resolve of a proxy ref to TAGGED appearance targets ("g:<geomAppId>" mesh | "o:<objAppId>"
-    // instance-object for ByBlock). (1) names the mesh; (2) names a DataObject/instance → its targets.
-    // NOT the layer tier — that's applied second, lower precedence, in BindWithPrecedence.
-    IReadOnlyList<string> DirectGeomKeys(string refAppId)
-    {
-      if (objectDisplayGeomKeys.TryGetValue(refAppId, out var og))
-      {
-        return og; // already tagged g:/o:
-      }
-      if (seenGeometryAppIds.Contains(refAppId))
-      {
-        return new[] { "g:" + refAppId };
-      }
-      return System.Array.Empty<string>();
-    }
-
-    // Bind nodes (material/colour) to geometry with CAD walk-up precedence: a DIRECT (mesh/object) ref
-    // claims a geometry first; a Layer ref then fills only geometry not already directly claimed — so an
-    // object's own colour overrides its layer's. Returns geomAppId → nodeK (deduped: one per mesh).
-    Dictionary<string, int> BindWithPrecedence(List<(int nodeK, List<string> refs)> proxies, out int skipped)
-    {
-      var byGeom = new Dictionary<string, int>(StringComparer.Ordinal);
-      var skip = 0;
-      foreach (var (nodeK, refs) in proxies)
-      {
-        foreach (var r in refs)
-        {
-          var direct = DirectGeomKeys(r);
-          if (direct.Count == 0 && !layerGeomKeys.ContainsKey(r))
-          {
-            skip++;
-          }
-          foreach (var gk in direct)
-          {
-            byGeom.TryAdd(gk, nodeK);
-          }
-        }
-      }
-      // Layer tier: a geometry can be enclosed by several nested layers (e.g. cladding under the specific
-      // "RAINSCREEN" layer, itself under a "WALL BUILDUP" grouping layer — both carry a ByLayer colour). The
-      // DEEPEST (most specific) layer must win, else the parent grouping's colour floods all its sub-layers
-      // (black "WALL BUILDUP" over tan "RAINSCREEN"). Collect the deepest candidate per geometry, then bind.
-      var layerCandidate = new Dictionary<string, (int depth, int nodeK)>(StringComparer.Ordinal);
-      foreach (var (nodeK, refs) in proxies)
-      {
-        foreach (var r in refs)
-        {
-          if (!layerGeomKeys.TryGetValue(r, out var lg))
-          {
-            continue;
-          }
-          var depth = layerDepth.GetValueOrDefault(r, 0);
-          foreach (var gk in lg)
-          {
-            if (byGeom.ContainsKey(gk))
-            {
-              continue; // a direct (mesh/object) ref already claimed it — beats any layer
-            }
-            if (!layerCandidate.TryGetValue(gk, out var cur) || depth > cur.depth)
-            {
-              layerCandidate[gk] = (depth, nodeK);
-            }
-          }
-        }
-      }
-      foreach (var (gk, (_, nodeK)) in layerCandidate)
-      {
-        byGeom[gk] = nodeK;
-      }
-      skipped = skip;
-      return byGeom;
-    }
-
-    // Definitions: members are either raw geometry → DEFINES (rel 4, → geometry), or nested instances →
-    // DEFINES_INSTANCE (rel 9, → INSTANCE node). The rel fixes the dst namespace (gap 2).
-    foreach (var def in GetBaseList(root, "instanceDefinitionProxies"))
-    {
-      if (def is not InstanceDefinitionProxy idp)
-      {
-        continue;
-      }
-      var defK = pipeline.AddDefinition(DefinitionKey(idp), idp.name);
-      stats.Definitions++;
-      int o = 0;
-      foreach (var memberAppId in idp.objects)
-      {
-        if (instanceNodeByAppId.TryGetValue(memberAppId, out var instK))
-        {
-          pipeline.DefinesInstance(defK, instK, o++);
-          stats.DefinesInstanceEdges++;
-        }
-        else if (seenGeometryAppIds.Contains(memberAppId))
-        {
-          pipeline.Defines(defK, pipeline.InternGeometryId(memberAppId), o++);
-          stats.DefinesEdges++;
-        }
-        else
-        {
-          // member geometry isn't in the bundle — skip rather than mint a phantom K.
-          stats.SkippedDefines++;
-        }
-      }
-    }
-
-    // Materials: HAS_MATERIAL(mesh → material), per-mesh, deduped with object>layer precedence. objects[]
-    // may name the mesh (BIM), the OBJECT (object-grained), or a Layer (ByLayer, resolved at produce).
-    var matProxies = new List<(int, List<string>)>();
-    var placeholderMatKs = new HashSet<int>();
-    foreach (var mat in GetBaseList(root, "renderMaterialProxies"))
-    {
-      if (mat is not RenderMaterialProxy rmp)
-      {
-        continue;
-      }
-      var v = rmp.value;
-      var matK = pipeline.AddMaterial(MaterialKey(rmp), v.diffuse, v.opacity, v.metalness, v.roughness);
-      matProxies.Add((matK, rmp.objects));
-      stats.Materials++;
-      // A pure-black diffuse is the CAD "no material / ByLayer" placeholder — must not override a real
-      // display colour (the loader applies HAS_MATERIAL over HAS_COLOR). Real materials (Rhino tans etc.)
-      // are kept; only the black placeholder yields to colour.
-      if ((v.diffuse & 0xFFFFFF) == 0)
-      {
-        placeholderMatKs.Add(matK);
-      }
-    }
-    var matBindings = BindWithPrecedence(matProxies, out var matSkipped);
-    stats.SkippedMaterial += matSkipped;
-
-    // Fallback tier: embedded per-mesh materials, for old-style sends that carry no root proxies (pre-proxy
-    // Revit embeds a RenderMaterial on each display mesh). A proxy claim wins; otherwise the mesh's own
-    // material binds directly. Deduped by material identity so 400 meshes sharing "Concrete - Cast In Situ"
-    // mint one MATERIAL node. NOT placeholder-eligible: an embedded black is an explicit assignment
-    // ("Black Glass"), not the CAD no-material sentinel.
-    var embeddedMatKs = new Dictionary<string, int>(StringComparer.Ordinal);
-    foreach (var (gAppId, rm) in embeddedMaterialByGeom)
-    {
-      var gk = "g:" + gAppId;
-      if (matBindings.ContainsKey(gk))
-      {
-        continue;
-      }
-      var key = rm.applicationId ?? "mat:" + (rm.id ?? rm.diffuse.ToString(CultureInfo.InvariantCulture));
-      if (!embeddedMatKs.TryGetValue(key, out var matK))
-      {
-        matK = pipeline.AddMaterial(key, rm.diffuse, rm.opacity, rm.metalness, rm.roughness);
-        embeddedMatKs[key] = matK;
-        stats.Materials++;
-      }
-      matBindings[gk] = matK;
-    }
-
-    // Colors: HAS_COLOR(mesh → colour), per-mesh — same 3-tier resolve + object>layer precedence.
-    var colProxies = new List<(int, List<string>)>();
-    foreach (var col in GetBaseList(root, "colorProxies"))
-    {
-      if (col is not ColorProxy cp)
-      {
-        continue;
-      }
-      colProxies.Add((pipeline.AddColor(cp.value), cp.objects));
-      stats.Colors++;
-    }
-    var colBindings = BindWithPrecedence(colProxies, out var colSkipped);
-    stats.SkippedColor += colSkipped;
-
-    // Emit. Targets are tagged "g:<geomAppId>" (mesh) or "o:<objAppId>" (instance object, ByBlock).
-    // MATERIAL (rel 5) is geometry-only; also drop a placeholder-black material on a mesh that has a real
-    // display colour so the colour wins.
-    foreach (var (target, matK) in matBindings)
-    {
-      if (target[0] != 'g')
-      {
-        continue; // instances don't carry a per-mesh material
-      }
-      if (placeholderMatKs.Contains(matK) && colBindings.ContainsKey(target))
-      {
-        continue;
-      }
-      pipeline.HasMaterial(pipeline.InternGeometryId(target[2..]), matK);
-      stats.HasMaterialEdges++;
-    }
-    // COLOUR (rel 6): GEOMETRY-targeted only. A geometry-K and an object-K share one dense numeric space, but
-    // the consumer keys HAS_COLOR by raw src into a SINGLE map (colors.get(geomK) ?? colors.get(objK)) — so an
-    // object-targeted colour and a geometry-targeted colour at the same numeric K COLLIDE (a window frame's def
-    // geometry picked up an unrelated plot-hatch instance's green). The earlier ByBlock fallback bound colours
-    // to the instance OBJECT (o:); now that def geometry carries its real SOURCE-LAYER colour on the geometry-K
-    // (see EmitDefinitionMember), that object fallback is redundant — drop it, killing the collision.
-    foreach (var (target, colK) in colBindings)
-    {
-      if (target[0] == 'o')
-      {
-        continue;
-      }
-      pipeline.HasColor(pipeline.InternGeometryId(target[2..]), colK);
-      stats.HasColorEdges++;
-    }
-
-    // Levels: ON_LEVEL(object → level). Generic read (no strong-typed LevelProxy in the SDK).
-    // Refs to elements absent from the export (e.g. Revit Rooms in a "sans rooms" send) are skipped.
-    foreach (var lvl in GetBaseList(root, "levelProxies"))
-    {
-      var members = lvl.GetMembers(DynamicBaseMemberType.Instance | DynamicBaseMemberType.Dynamic);
-      var (name, elevation) = ReadLevelValue(members);
-      var lvlK = pipeline.AddLevel(LevelKey(lvl, name), name, elevation);
-      stats.Levels++;
-      foreach (var objAppId in AsStringList(members.GetValueOrDefault("objects")))
-      {
-        if (!seenObjectAppIds.Contains(objAppId))
-        {
-          stats.SkippedLevel++;
-          continue;
-        }
-        pipeline.OnLevel(pipeline.InternObject(objAppId), lvlK);
-        stats.OnLevelEdges++;
-      }
-    }
-  }
-
-  // ── traversal ────────────────────────────────────────────────────────────────────
-
-  // Members that hold geometry or proxy data, NOT child objects — never descended as containers.
-  // (`displayValue` is the object's own geometry, handled in EmitObject; the proxy arrays live on the root.)
-  private static readonly HashSet<string> s_nonContainerMembers = new(StringComparer.Ordinal)
-  {
-    "displayValue",
-    "@displayValue",
-    "baseCurves",
-    "@baseCurves",
-    "instanceDefinitionProxies",
-    "@instanceDefinitionProxies",
-    "renderMaterialProxies",
-    "@renderMaterialProxies",
-    "colorProxies",
-    "@colorProxies",
-    "levelProxies",
-    "@levelProxies",
-    "groupProxies",
-    "@groupProxies",
-    // non-scene metadata containers (not renderable model objects)
-    "cameras",
-    "@cameras",
-    "views",
-    "@views",
-  };
-
-  // Yields every LEAF reachable through any container member, with its parent. A leaf is atomic
-  // type-agnostically (SOT §1) — DataObject, raw geometry, or a bare InstanceProxy placement. Descent
-  // follows `elements` AND any other Base-list member (e.g. Civil3D `@mainBaselineFeatureLines`, `@regions`)
-  // except geometry/proxy members — so awkwardly-nested geometry is reached, not just `elements`.
-  private static IEnumerable<(Base obj, Base? parent)> TraverseAtomics(Base root)
-  {
-    var stack = new Stack<(Base node, Base? parent)>();
-    var visited = new HashSet<string>(StringComparer.Ordinal);
-    stack.Push((root, null));
-    while (stack.Count > 0)
-    {
-      var (node, parent) = stack.Pop();
-      if (node.id is { } id && !visited.Add(id))
-      {
-        continue; // already walked (shared reference) — avoid re-descending
-      }
-      var isContainer = node is Collection || ReferenceEquals(node, root);
-      if (!isContainer && !IsValueOrDefinitionProxy(node))
-      {
-        yield return (node, parent);
-      }
-      // Geometry is TERMINAL — never descend into it. Its list-members (Polycurve `segments`, Mesh
-      // `vertices`/`faces`, etc.) are internal components, not child scene objects; descending would extract
-      // them as spurious fragment-objects (and they'd miss layer-colour resolution).
-      if (IsGeometry(node))
-      {
-        continue;
-      }
-      foreach (var child in ChildContainers(node))
-      {
-        stack.Push((child, node));
-      }
-    }
-  }
-
-  // layer/collection appId → the display-geometry appIds of every object beneath it (any depth). Lets a
-  // ByLayer material/colour proxy (which references a Layer) bind to that layer's objects' meshes — the
-  // CAD "walk up to the layer's colour" resolved once at produce, keeping the envelope flat.
-  // Scene-tree containers → COLLECTION nodes (root excluded), parent-chained. Generic over source: Rhino/CAD
-  // `Layer` collections, ETABS story→category, Tekla category — any Collection in the tree. The parent chain
-  // (COLLECTION.def_ref) reconstructs the source hierarchy that flattening loses; the subtype tag
-  // (collectionType ?? speckle_type leaf) lets the loader distinguish layer / category / story.
-  private static Dictionary<string, int> BuildCollectionNodes(Base root, ObjectsArtifactPipeline pipeline, Stats stats)
-  {
-    var map = new Dictionary<string, int>(StringComparer.Ordinal);
-    void Walk(Base node, int? parentK)
-    {
-      foreach (var child in ChildContainers(node))
-      {
-        if (child is not Collection col)
-        {
-          continue;
-        }
-        var appId = Aid(child);
-        if (!map.TryGetValue(appId, out var k))
-        {
-          k = pipeline.AddCollection(CollectionKey(col), col.name, parentK, CollectionSubtype(col));
-          map[appId] = k;
-          stats.Collections++;
-        }
-        Walk(child, k);
-      }
-    }
-    Walk(root, null); // root itself is NOT a node — its child collections are top-level (parentK = null)
-    return map;
-  }
-
-  private static string CollectionKey(Collection col) => col.applicationId ?? "coll:" + col.id;
-
-  // Subtype tag for the loader: prefer the source collectionType ("layer" on SketchUp), else the speckle_type
-  // leaf ("Layer" on Rhino/AutoCAD/Civil3D, "Collection" on Revit/Tekla/ETABS category grouping).
-  private static string CollectionSubtype(Collection col)
-  {
-    var members = col.GetMembers(DynamicBaseMemberType.Instance | DynamicBaseMemberType.Dynamic);
-    var ct = members.GetValueOrDefault("collectionType") as string;
-    return !string.IsNullOrEmpty(ct) ? ct : col.speckle_type.Split('.')[^1];
-  }
-
-  private static Dictionary<string, List<string>> BuildLayerGeomKeys(
-    Base root,
-    Dictionary<string, List<string>> objectDisplayGeomKeys,
-    out Dictionary<string, int> layerDepth
-  )
-  {
-    var result = new Dictionary<string, List<string>>(StringComparer.Ordinal);
-    var depths = new Dictionary<string, int>(StringComparer.Ordinal);
-    void Walk(Base node, int depth)
-    {
-      foreach (var child in ChildContainers(node))
-      {
-        if (child is Collection)
-        {
-          var geoms = new List<string>();
-          CollectDescendantGeom(child, objectDisplayGeomKeys, geoms);
-          if (geoms.Count > 0)
-          {
-            result[Aid(child)] = geoms;
-            depths[Aid(child)] = depth;
-          }
-          Walk(child, depth + 1);
-        }
-      }
-    }
-    Walk(root, 0);
-    layerDepth = depths;
-    return result;
-  }
-
-  private static void CollectDescendantGeom(
-    Base collection,
-    Dictionary<string, List<string>> objectDisplayGeomKeys,
-    List<string> acc
-  )
-  {
-    foreach (var child in ChildContainers(collection))
-    {
-      if (child is Collection)
-      {
-        CollectDescendantGeom(child, objectDisplayGeomKeys, acc);
-      }
-      else if (objectDisplayGeomKeys.TryGetValue(Aid(child), out var gks))
-      {
-        acc.AddRange(gks);
-      }
-    }
-  }
-
-  // All child Base objects under `node`'s container members (every Base-list member except the
-  // geometry/proxy members in s_nonContainerMembers).
-  private static IEnumerable<Base> ChildContainers(Base node)
-  {
-    var members = node.GetMembers(DynamicBaseMemberType.Instance | DynamicBaseMemberType.Dynamic);
-    foreach (var kv in members)
-    {
-      if (s_nonContainerMembers.Contains(kv.Key))
-      {
-        continue;
-      }
-      if (kv.Value is System.Collections.IEnumerable seq and not string)
-      {
-        foreach (var item in seq)
-        {
-          if (item is Base b)
-          {
-            yield return b;
-          }
-        }
-      }
-    }
-  }
-
-  // ── property extraction ──────────────────────────────────────────────────────────
-
-  private static (
-    IReadOnlyDictionary<string, object?> props,
-    IEnumerable<KeyValuePair<string, object?>> rootScalars,
-    string? typeKey
-  ) ExtractProperties(Base obj)
-  {
-    IReadOnlyDictionary<string, object?> props = obj is DataObject dobj
-      ? dobj.properties
-      : obj.GetMembers(DynamicBaseMemberType.Instance | DynamicBaseMemberType.Dynamic);
-
-    var members = obj.GetMembers(DynamicBaseMemberType.Instance | DynamicBaseMemberType.Dynamic);
-    // Root-level scalar fields indexed outside `properties` (mirrors the SDK's EavExtraction
-    // s_rootScalarFields). `level` is a top-level member on Revit DataObjects = the level NAME
-    // ("GRND" etc.) — it lives at the top level, NOT under properties, so it must be listed here
-    // explicitly or it is dropped (the deep `properties.Parameters…Level` rows are unrelated params).
-    var rootScalars = new List<KeyValuePair<string, object?>>
-    {
-      new("speckle_type", obj.speckle_type),
-      new("name", members.GetValueOrDefault("name")),
-      new("units", members.GetValueOrDefault("units")),
-      new("category", members.GetValueOrDefault("category")),
-      new("family", members.GetValueOrDefault("family")),
-      new("type", members.GetValueOrDefault("type")),
-      new("level", members.GetValueOrDefault("level")),
-    };
-
-    var typeKey =
-      members.GetValueOrDefault("typeId") as string ?? (props.TryGetValue("typeId", out var tk) ? tk as string : null);
-
-    return (props, rootScalars, typeKey);
-  }
-
-  private static (string? name, double elevation) ReadLevelValue(IReadOnlyDictionary<string, object?> members)
-  {
-    if (members.GetValueOrDefault("value") is Base valueBase)
-    {
-      var vm = valueBase.GetMembers(DynamicBaseMemberType.Instance | DynamicBaseMemberType.Dynamic);
-      return (vm.GetValueOrDefault("name") as string, ToDouble(vm.GetValueOrDefault("elevation")));
-    }
-    if (members.GetValueOrDefault("value") is IReadOnlyDictionary<string, object?> vd)
-    {
-      return (vd.GetValueOrDefault("name") as string, ToDouble(vd.GetValueOrDefault("elevation")));
-    }
-    return (members.GetValueOrDefault("name") as string, ToDouble(members.GetValueOrDefault("elevation")));
-  }
-
-  // ── id keying (everything is applicationId-keyed; null → stable per-object key) ───
-
-  private static string Aid(Base b) => b.applicationId ?? "spk:" + b.id;
-
-  private static string GeometryKey(Base mesh) => mesh.applicationId ?? "spk:" + mesh.id;
-
-  private static string DefinitionKey(InstanceDefinitionProxy idp) => idp.applicationId ?? idp.name;
-
-  private static string MaterialKey(RenderMaterialProxy rmp) =>
-    rmp.applicationId ?? rmp.value.applicationId ?? "mat:" + rmp.value.diffuse.ToString(CultureInfo.InvariantCulture);
-
-  private static string LevelKey(Base lvl, string? name) => lvl.applicationId ?? "lvl:" + (name ?? lvl.id);
-
-  // ── helpers ──────────────────────────────────────────────────────────────────────
-
-  // Value/definition proxies that are NEVER atomic leaves (they live on the root's dynamic members).
-  // Note: InstanceProxy is deliberately excluded — a bare instance placement CAN be an atomic leaf.
-  private static bool IsValueOrDefinitionProxy(Base b) =>
-    b is InstanceDefinitionProxy or RenderMaterialProxy or ColorProxy or GroupProxy;
-
-  private static bool IsGeometry(Base b) => b.speckle_type.StartsWith("Objects.Geometry.", StringComparison.Ordinal);
-
-  // The inline `renderMaterial` member old-style sends attach to a display mesh (or its parent element) —
-  // dynamic on Mesh, so read via members (typed OR `@`-prefixed detached key).
-  private static RenderMaterial? ReadEmbeddedMaterial(Base host)
-  {
-    var members = host.GetMembers(DynamicBaseMemberType.Instance | DynamicBaseMemberType.Dynamic);
-    return (members.GetValueOrDefault("renderMaterial") ?? members.GetValueOrDefault("@renderMaterial"))
-      as RenderMaterial;
-  }
-
-  private static IEnumerable<Base> GetBaseList(Base b, string key)
-  {
-    var members = b.GetMembers(DynamicBaseMemberType.Instance | DynamicBaseMemberType.Dynamic);
-    // A detached list may live under the typed key OR the `@`-prefixed dynamic key (older connectors /
-    // app.speckle.systems keep `@elements`/`@displayValue` as dynamic members the deserializer didn't map
-    // onto the typed property). Take whichever is a non-empty Base list.
-    var raw = NonEmpty(members.GetValueOrDefault(key)) ?? NonEmpty(members.GetValueOrDefault("@" + key));
-    if (raw is System.Collections.IEnumerable seq and not string)
-    {
-      foreach (var item in seq)
-      {
-        if (item is Base bs)
-        {
-          yield return bs;
-        }
-      }
-    }
-  }
-
-  // Returns the value only if it's a non-empty enumerable (so an empty typed `elements` falls through to
-  // the `@elements` dynamic member).
-  private static object? NonEmpty(object? v) => v is System.Collections.ICollection c && c.Count == 0 ? null : v;
-
-  private static IEnumerable<string> AsStringList(object? raw)
-  {
-    if (raw is System.Collections.IEnumerable seq and not string)
-    {
-      foreach (var item in seq)
-      {
-        if (item is string s)
-        {
-          yield return s;
-        }
-      }
-    }
-  }
-
-  private static double ToDouble(object? o) =>
-    o switch
-    {
-      double d => d,
-      float f => f,
-      int i => i,
-      long l => l,
-      _ => 0.0,
-    };
-
-  private static double[] Flatten(Matrix4x4 m) =>
-    new[]
-    {
-      m.M11,
-      m.M12,
-      m.M13,
-      m.M14,
-      m.M21,
-      m.M22,
-      m.M23,
-      m.M24,
-      m.M31,
-      m.M32,
-      m.M33,
-      m.M34,
-      m.M41,
-      m.M42,
-      m.M43,
-      m.M44,
-    };
 }
