+++ src/Speckle.Sdk.Artifacts.Harness/Speckle.Sdk.Artifacts.Harness.csproj
@@ -4,14 +4,12 @@
     <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
-    <!--    <TestProjectAnalyserRules>true</TestProjectAnalyserRules>-->
-    <!--    &lt;!&ndash; Correctness-iteration harness: relax the strictest analyzers so we can move fast.-->
-    <!--         The driver gets promoted to src/Speckle.Objects (full rules) once proven. &ndash;&gt;-->
     <NoWarn>$(NoWarn);CA2007</NoWarn>
   </PropertyGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.Logging.Console" />
     <PackageReference Include="System.CommandLine" />
+    <PackageReference Include="DuckDB.NET.Data.Full" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\src\Speckle.Objects\Speckle.Objects.csproj" />
